requirements.txt
asgiref==3.3.4
Django==2.2.24
django-admin-numeric-filter==0.1.6
django-material-admin==1.8.5
pytz==2021.1
sqlparse==0.4.1
